+++ src/config.h
@@ -254,6 +254,7 @@ typedef struct {
   string word_chars_excl;
   colour ime_cursor_colour;
   colour_pair ansi_colours[16];
+  int max_image_size;
   wstring sixel_clip_char;
   bool short_long_opts;
   bool bold_as_special;